+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - **Kubernetes:** Fixed backend config credentials to support all available AWS ARN partitions (Issue [#913](https://github.com/NetApp/trident/issues/913)).
 - **Kubernetes:** Added option to disable the auto configurator reconciliation in the Trident operator (Issue [#924](https://github.com/NetApp/trident/issues/924)).
 - **Kubernetes:** Added securityContext for csi-resizer container (Issue [#976](https://github.com/NetApp/trident/issues/976)).
+- **Kubernetes:** Fixed issue where iSCSI self-healing may unsafely log out of stale iSCSI sessions when a portal is inaccessible (Issue [#961](https://github.com/NetApp/trident/issues/961)).
 - Fixed Zonal Flex pools for GCNV driver.
 
 **Enhancements:**